Write Stata data to pre-formatted Excel file
Load some demo data, add rating colors and trends, make a copy of the 
Excel template, and then write the Stata data to the Excel file.

In the Excel file, conditional formatting is kept intact. However, it 
turns out that normal formatting is lost when Stata writes data to the 
cells UNLESS those cells already had text in them before.

That is not very useful for us because we do not know ahead of time how 
many columns we will need. Two alternatives:
1) Use conditional formatting instead of normal formatting also for the 
headers.
2) Style the Excel file with Stata's excelput command.